+++ app/Http/Controllers/Admin/ProductController.php
@@ -47,12 +47,10 @@ class ProductController extends Controller
         if($req->hasFile('image')){
             $image = $req->file('image');
             $filename = time() . '.' . $image->getClientOriginalExtension();
-            Image::make($image)->save(public_path('photos/'.$filename));
+            Image::make($image)->save(public_path('images/upload/'.$filename));
 
-            $product->image = ('photos/'.$filename);
+            $product->image = ('images/upload/'.$filename);
 
-        } else{
-            $product->image = ('photos/avatar5.png');
         }
         $product->save();
 
@@ -85,12 +83,14 @@ class ProductController extends Controller
         $product->meta_key = $req['meta_key'];
         $product->meta_des = $req['meta_des'];
         if($req->hasFile('image')){
-            unlink($product->image);
+            if(file_exists($product->image))
+            {
+                unlink($product->image);
+            }
             $image = $req->file('image');
             $filename = time() . '.' . $image->getClientOriginalExtension();
-            Image::make($image)->save(public_path('photos/'.$filename));
-            $product->image = ('photos/'.$filename);
-
+            Image::make($image)->save(public_path('images/upload/'.$filename));
+            $product->image = ('images/upload/'.$filename);
         }
         $validatedData = $req->validate([
             'name' => 'required|unique:products,name,' .$product->id,
@@ -118,7 +118,7 @@ class ProductController extends Controller
     {
         $id_cate_product = $req->cate_product;
         if($id_cate_product == 0){
-            return back();
+            return redirect()->route('admin.product.index');
         }
         session()->put('id',$id_cate_product);
         $data=Cate_product::select('id','name','parent_id')->get()->toArray();
